🦄 refactor(kernel): seperate arch-specific functions
Seperate arch-specific functions from the kernel to decouple.
