+++ include/sus/arch.h
@@ -1,0 +1,25 @@
+/**
+ * @file arch.h
+ * @author theflysong (song_of_the_fly@163.com)
+ * @brief 架构相关函数声明
+ *        这部分包含与特定架构相关的初始化和功能实现
+ * @version alpha-1.0.0
+ * @date 2025-11-21
+ * 
+ * @copyright Copyright (c) 2025
+ * 
+ */
+
+#pragma once
+
+/**
+ * @brief 架构特定初始化
+ * 
+ */
+void arch_init(void);
+
+/**
+ * @brief 架构特定收尾操作
+ * 
+ */
+void arch_terminate(void);
