+++ kernel/arch/riscv64/include.mk
@@ -1,1 +1,2 @@
 objects += ./arch/riscv64/entry.o
+objects += ./arch/riscv64/setup.o
